+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+
+### Added
+
+- A `fetchStatusCodeClasses` function, to fetch the HTTP status code classes.
+
 [0.1.1] - 2025-11-26
 --------------------
 
@@ -20,5 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[UNRELEASED]: https://github.com/jbenner-radham/node-fetch-status-codes/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/jbenner-radham/node-fetch-status-codes/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/jbenner-radham/node-fetch-status-codes/releases/tag/v0.1.0